Update README.md
This is my web application placeholder. Here, I have completed the demo of my web application. I am having an issue with my backend web application. The problem is regarding Integrating it with Gmaps, which I can't do because there is no Application that can be used, although I have used APIs like Google Map Developer and Openstreet Map for it.

The parking feature allows users to select parking spaces based on their vehicle type (2-wheeler, 3-wheeler, or 4-wheeler). Users are presented with a map displaying parking spots categorized as available (green) or occupied (red). The system enables users to reserve a spot by interacting with the map. This feature is dynamically updated to reflect real-time availability, ensuring accurate information for the users.